checkTotal() bug fix

+++ ClimsoftVer4/ClimsoftVer4/ucrFormDaily2.vb
@@ -339,8 +339,8 @@ Public Class ucrFormDaily2
 
     Public Function checkTotal() As Boolean
         Dim bValueCorrect As Boolean = False
-        Dim elemTotal As Integer = 0
-        Dim expectedTotal As Integer
+        Dim elemTotal As Decimal = 0
+        Dim expectedTotal As Decimal
 
         If bTotalRequired Then
             If ucrInputTotal.IsEmpty AndAlso Not IsValuesEmpty() Then
